Struct tfdeploy::Model

pub struct Model { /* fields omitted */ }

Model is Tfdeploy workhouse. It wraps a protobuf tensorflow model, and runs the inference interpreter.

Methods

impl Model

pub fn new(graph: GraphDef) -> Result<Model>

pub fn node_id_by_name(&self, name: &str) -> Result<usize>

pub fn state(&self) -> ModelState

pub fn for_path<P: AsRef<Path>>(p: P) -> Result<Model>

Load a Tensorflow protobul model from a file.

pub fn for_reader<R: Read>(r: R) -> Result<Model>

Load a Tfdeploy model from a reader.

pub fn graphdef_for_reader<R: Read>(r: R) -> Result<GraphDef>

Load a Tensorflow protobuf graph def from a reader.

pub fn graphdef_for_path<P: AsRef<Path>>(p: P) -> Result<GraphDef>

Load a Tensorflow protobuf graph def from a path

pub fn node_names(&self) -> Vec<&str>

pub fn get_node(&self, name: &str) -> Result<&Node>

Build a tfdeploy Node by name.

pub fn plan_for_one(&self, node: usize) -> Result<Plan>

pub fn run(
&self,
inputs: Vec<(usize, Matrix)>,
output: usize
) -> Result<Vec<Matrix>>

pub fn nodes(&self) -> &[Node]

pub fn run_with_names(
&self,
inputs: Vec<(&str, Matrix)>,
output: &str
) -> Result<Vec<Matrix>>